+++ src/main/java/com/wedevol/springdatacouchbase/core/dao/UserRepository.java
@@ -37,7 +37,7 @@ public interface UserRepository extends CrudRepository<UserDoc, String> {
   @Query("SELECT COUNT(u.id) AS c FROM users u WHERE u.type = 'com.wedevol.springdatacouchbase.core.dao.doc.UserDoc'")
   Integer countAll();
 
-  // Raw N1QL query hat projects some attributes (not all) using like operator and a parameterized
+  // Raw N1QL query that projects some attributes (not all) using like operator and a parameterized
   // variable
   @Query("SELECT u.id, u.name, META(u).id AS _ID, META(u).cas AS _CAS FROM users u WHERE u.type = 'com.wedevol.springdatacouchbase.core.dao.doc.UserDoc' AND LOWER(u.name) LIKE '%' || $name || '%'")
   List<UserBasicDoc> findUsersWithName(@Param("name") String name);