+++ src/main/java/com/wedevol/springdatacouchbase/core/dao/doc/UserDoc.java
@@ -7,7 +7,7 @@ import com.couchbase.client.java.repository.annotation.Field;
 import com.couchbase.client.java.repository.annotation.Id;
 
 /**
- * Represents a user doc from Couchbase. The doc key is generated using a Couchbase atomic counter.
+ * Represents a user doc from Couchbase. The doc key is generated using a Couchbase atomic counter with manual prefix.
  *
  * @author Charz++
  */
@@ -107,12 +107,7 @@ public class UserDoc implements Serializable {
   public int hashCode() {
     final int prime = 31;
     int result = 1;
-    result = prime * result + ((age == null) ? 0 : age.hashCode());
-    result = prime * result + ((email == null) ? 0 : email.hashCode());
     result = prime * result + ((id == null) ? 0 : id.hashCode());
-    result = prime * result + ((key == null) ? 0 : key.hashCode());
-    result = prime * result + ((name == null) ? 0 : name.hashCode());
-    result = prime * result + ((nicknames == null) ? 0 : nicknames.hashCode());
     return result;
   }
 
@@ -125,35 +120,10 @@ public class UserDoc implements Serializable {
     if (getClass() != obj.getClass())
       return false;
     UserDoc other = (UserDoc) obj;
-    if (age == null) {
-      if (other.age != null)
-        return false;
-    } else if (!age.equals(other.age))
-      return false;
-    if (email == null) {
-      if (other.email != null)
-        return false;
-    } else if (!email.equals(other.email))
-      return false;
     if (id == null) {
       if (other.id != null)
         return false;
     } else if (!id.equals(other.id))
-      return false;
-    if (key == null) {
-      if (other.key != null)
-        return false;
-    } else if (!key.equals(other.key))
-      return false;
-    if (name == null) {
-      if (other.name != null)
-        return false;
-    } else if (!name.equals(other.name))
-      return false;
-    if (nicknames == null) {
-      if (other.nicknames != null)
-        return false;
-    } else if (!nicknames.equals(other.nicknames))
       return false;
     return true;
   }