+++ src/main/java/com/wedevol/springdatacouchbase/core/service/PhoneService.java
@@ -1,0 +1,16 @@
+package com.wedevol.springdatacouchbase.core.service;
+
+import com.wedevol.springdatacouchbase.core.dao.doc.PhoneDoc;
+
+/**
+ * Interface for the Phone Service Implementation
+ *
+ * @author Charz++
+ */
+
+public interface PhoneService {
+
+  PhoneDoc findById(String id);
+
+  PhoneDoc create(PhoneDoc phone);
+}