+++ src/main/java/com/wedevol/springdatacouchbase/core/service/impl/PlaceServiceImpl.java
@@ -1,6 +1,8 @@
 package com.wedevol.springdatacouchbase.core.service.impl;
 
 import java.util.Optional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.wedevol.springdatacouchbase.core.dao.PlaceCounterRepository;
@@ -11,14 +13,15 @@ import com.wedevol.springdatacouchbase.core.exception.ErrorType;
 import com.wedevol.springdatacouchbase.core.service.PlaceService;
 
 /**
- * Service that manages the creation of a Place to exemplify the use of multiple templates and
- * repositories.
+ * Service that manages the creation of a Place to exemplify the use of multiple templates and repositories.
  *
  * @author Charz++
  */
 
 @Service
 public class PlaceServiceImpl implements PlaceService {
+
+  protected static final Logger logger = LoggerFactory.getLogger(PlaceServiceImpl.class);
 
   @Autowired
   private PlaceRepository placeRepo;
@@ -28,14 +31,16 @@ public class PlaceServiceImpl implements PlaceService {
 
   @Override
   public PlaceDoc findById(Long id) {
-    final Optional<PlaceDoc> placeObj = placeRepo.findById(PlaceDoc.getKeyFor(id));
+    Optional<PlaceDoc> placeObj = placeRepo.findById(PlaceDoc.getKeyFor(id));
     return placeObj.orElseThrow(() -> new ApiException(ErrorType.PLACE_NOT_FOUND));
   }
 
   @Override
   public PlaceDoc create(PlaceDoc place) {
     place.setId(placeCounterRepo.counter()); // internally we set the key with that id
-    return placeRepo.save(place);
+    PlaceDoc placeFromDb = placeRepo.save(place);
+    logger.info("place key: {}", placeFromDb.getKey());
+    return placeFromDb;
   }
 
 }