+++ src/main/java/com/wedevol/springdatacouchbase/core/service/impl/ProductServiceImpl.java
@@ -1,6 +1,8 @@
 package com.wedevol.springdatacouchbase.core.service.impl;
 
 import java.util.Optional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.couchbase.config.BeanNames;
@@ -12,8 +14,9 @@ import com.wedevol.springdatacouchbase.core.exception.ErrorType;
 import com.wedevol.springdatacouchbase.core.service.ProductService;
 
 /**
- * Service that manages the creation of a Product to exemplify the use of unique number (uuid) for
- * key generation.
+ * Service that manages the creation of a Product to exemplify the use of unique number (Couchbase UUID) for key
+ * generation and the usage of the template methods (not the CRUD repository). NOTE: As of 2019 I could not save the
+ * auto-generated unique Couchbase key inside the doc.
  *
  * @author Charz++
  */
@@ -21,20 +24,23 @@ import com.wedevol.springdatacouchbase.core.service.ProductService;
 @Service
 public class ProductServiceImpl implements ProductService {
 
-  // add the qualifier in case you have multiple buckets in your configuration otherwise remove it
+  protected static final Logger logger = LoggerFactory.getLogger(ProductServiceImpl.class);
+
+  // NOTE: add the qualifier in case you have multiple buckets in your configuration otherwise remove it
   @Autowired
   @Qualifier(BeanNames.COUCHBASE_TEMPLATE)
-  private CouchbaseTemplate template;
+  private CouchbaseTemplate defaultTemplate;
 
   @Override
-  public ProductDoc findById(String id) {
-    final Optional<ProductDoc> productObj = Optional.ofNullable(template.findById(id, ProductDoc.class));
+  public ProductDoc findById(String id) { // this is the key
+    Optional<ProductDoc> productObj = Optional.ofNullable(defaultTemplate.findById(id, ProductDoc.class));
     return productObj.orElseThrow(() -> new ApiException(ErrorType.PRODUCT_NOT_FOUND));
   }
 
   @Override
   public void create(ProductDoc product) {
-    template.insert(product);
+    defaultTemplate.insert(product);
+    logger.info("product key: {}", product.getId());
   }
 
 }