+++ src/main/java/com/wedevol/springdatacouchbase/core/util/Util.java
@@ -1,6 +1,7 @@
 package com.wedevol.springdatacouchbase.core.util;
 
 import java.util.List;
+import java.util.UUID;
 
 /**
  * Util class for constants and generic methods
@@ -16,6 +17,10 @@ public class Util {
 
   public static boolean isNullOrEmpty(List<?> element) {
     return element == null || element.isEmpty();
+  }
+  
+  public static String uuid() {
+    return UUID.randomUUID().toString();
   }
 
 }